+++ lib/challah/user/password.rb
@@ -6,6 +6,11 @@ module Challah::User
       self.password_confirmation = new_password
     end
 
+    def password_provider?
+      return true if @password_updated or @username_updated
+      !!providers[:password]
+    end
+
     # Set a password for this user
     def password=(value)
       if value.to_s.blank?
@@ -25,6 +30,15 @@ module Challah::User
     # Was the password updated
     def password_changed?
       !!@password
+    end
+
+    def username
+      @username ||= password_provider? ? password_provider[:uid] : ''
+    end
+
+    def username=(value)
+      @username_updated = true
+      @username = value.to_s.strip.downcase
     end
   end
 end